Fix multiple long piston bugs

+++ src/main/java/ca/fxco/pistonlib/blocks/pistons/basePiston/BasicPistonHeadBlock.java
@@ -121,9 +121,11 @@ public class BasicPistonHeadBlock extends DirectionalBlock {
         if (!newState.is(this)) {
             super.onRemove(state, level, pos, newState, movedByPiston);
 
-            BlockPos behindPos = pos.relative(state.getValue(FACING).getOpposite());
+            Direction facing = state.getValue(FACING);
+            BlockPos behindPos = pos.relative(facing.getOpposite());
 
-            if (this.isFittingBase(state, level.getBlockState(behindPos))) {
+            if (this.isFittingBase(state, level.getBlockState(behindPos)) &&
+                    !(newState.is(this.family.getArm()) && newState.getValue(FACING) == facing)) {
                 level.destroyBlock(behindPos, true);
             }
         }
